requirements.txt
arxiv>=1.4.8
pinecone-client>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=8.0.0
pytest-mock>=3.14.0
# Added for development and testing
pytest>=8.0.0  # For running unit tests
pytest-mock>=3.14.0  # For mocking in tests 